move group selection to component remove unused dep
