+++ package.json
@@ -37,7 +37,6 @@
     "vee-validate": "^4.13.2",
     "vite-plugin-html": "^3.2.2",
     "vue": "^3.5.12",
-    "vue-draggable-next": "^2.2.1",
     "vue-loading-overlay": "^6.0.6",
     "vue-matomo": "^4.2.0",
     "vue-multiselect": "^3.1.0",
