+++ src/components/forms/dataset/DatasetPropertiesFields.vue
@@ -14,9 +14,6 @@ import { useTopicsConf } from '@/utils/config'
 import DatasetPropertiesTextFields from './DatasetPropertiesTextFields.vue'
 import SelectDataset from './SelectDataset.vue'
 
-import Multiselect from '@vueform/multiselect'
-import '@vueform/multiselect/themes/default.css'
-
 const emit = defineEmits(['updateValidation', 'update:datasetProperties'])
 
 const datasetProperties = defineModel({
@@ -93,8 +90,6 @@ const onSelectDataset = (value: DatasetV2 | undefined) => {
   }
 }
 
-const groupOptions = ref(Array.from(datasetsGroups.value, ([key]) => key))
-
 watch(
   isValidDataset,
   (newValue) => {
@@ -142,29 +137,11 @@ onMounted(() => {
       Rechercher ou créer un regroupement. Un regroupement contient un ou
       plusieurs jeux de données.
     </p>
-    <Multiselect
-      id="input-regroupement"
-      v-model="datasetProperties.group"
-      :options="groupOptions"
-      :searchable="true"
-      :limit="5"
-      :strict="false"
-      no-options-text="Il n'y a pas encore de regroupement dans ce bouquet."
-      no-results-text="Aucun regroupement existant."
-      :create-option="true"
-      name="select"
-      placeholder=""
-      :aria="{
-        'aria-describedby': 'regroupement-description'
-      }"
-    >
-      <template #option="{ option }">
-        <p v-if="option.__CREATE__">
-          Ajouter "{{ option.label }}" comme regroupement
-        </p>
-        <p v-else>{{ option.label }}</p>
-      </template>
-    </Multiselect>
+
+    <SelectTopicGroup
+      v-model:properties-model="datasetProperties"
+      v-model:groups-model="datasetsGroups"
+    />
   </div>
   <div
     v-if="!selectedDataset && topicsDatasetEditorialization"
